Convert datepicker wrapper to ES6 module
