+++ ui/legacy.js
@@ -11,7 +11,6 @@ import './legacyHelpers';
 import '../app/assets/javascripts/student_profile/quad_converter.jsx';
 import '../app/assets/javascripts/student_profile/provided_by_educator_dropdown.jsx';
 import '../app/assets/javascripts/student_profile/profile_chart_settings.jsx';
-import '../app/assets/javascripts/student_profile/datepicker.jsx';
 import '../app/assets/javascripts/student_profile/bar_chart_sparkline.jsx';
 import '../app/assets/javascripts/student_profile/sparkline.jsx';
 import '../app/assets/javascripts/student_profile/academic_summary.jsx';
